\name{simDat5}
\alias{simDat5}
\docType{data}
\title{
Simulated Data
}
\description{
This dataset is used in several documentation files to illustrate the
different functionalities of the package.
}
\usage{data("simDat5")}
\format{
  A data frame with 300 observations on the following 16 variables. 
  \describe{
    \item{\code{X1}}{Continuous Covariate}
    \item{\code{X2}}{Continuous Covariate}
    \item{\code{Z}}{Treatment indicator.}
    \item{\code{Y}}{Observed outcome.}
    \item{\code{Y1}}{Outcome for the treated (not observed in
      practice).}      
    \item{\code{Y0}}{Outcome for the nontreated (not observed in
      practice).}
  }
}

\keyword{datasets}
